Выделит логику фильма в отдельный презентер
